backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
ecdsa==0.19.0
fastapi==0.110.3
filelock==3.14.0
h11==0.14.0
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
passlib==1.7.4
pipenv==2023.12.1
platformdirs==4.2.1
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.7.1
pydantic_core==2.18.2
pypdf==4.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing==3.7.4.3
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.1
polygon-api-client==1.13.7
python-dateutil==2.9.0.post0
